20-02-2025-ChatBot/requirements.txt
langchain_ollama
langchain_chroma
langchain_core
python-dotenv
streamlit